clean up SN76496 cores (MAME/Maxim)
changed int/INT32 to the smallest possible value type:
- UINT8 for flags
- UINT16 for register data
- UINT32 for noise masks
also removed unused parameters/variables from SN76489 Maxim core

+++ emu/cores/sn76496.h
@@ -9,12 +9,12 @@ void sn76496_write_reg(void *chip, UINT8 offset, UINT8 data);
 void sn76496_stereo_w(void *chip, UINT8 offset, UINT8 data);
 
 void SN76496Update(void *param, UINT32 samples, DEV_SMPL** outputs);
-UINT32 sn76496_start(void **chip, UINT32 clock, int shiftregwidth, int noisetaps,
-					int negate, int stereo, int clockdivider, int freq0);
+UINT32 sn76496_start(void **chip, UINT32 clock, UINT8 shiftregwidth, UINT16 noisetaps,
+					UINT8 negate, UINT8 stereo, UINT8 clockdivider, UINT8 freq0);
 void sn76496_connect_t6w28(void *noisechip, void *tonechip);
 void sn76496_shutdown(void *chip);
 void sn76496_reset(void *chip);
-void sn76496_freq_limiter(void* chip, int sample_rate);
+void sn76496_freq_limiter(void* chip, UINT32 sample_rate);
 void sn76496_set_mutemask(void *chip, UINT32 MuteMask);
 
 #endif	// __SN76496_H__
